+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2023-08-13
+
 ### Added
 - Implement the auto capitalization. [(#56)](https://github.com/pythonbrad/clafrica/pull/56)